+++ actions/create_client.py
@@ -2,5 +2,13 @@ from lib import action
 
 
 class KeycloakClientCreateAction(action.KeycloakBaseAction):
-    def run(self, name, client_id, redirect_uris, protocol, public_client, direct_access_grants):
-        self.keycloak_admin.create_client(name=name, client_id=client_id, redirect_uris=redirect_uris, protocol=protocol, public_client=public_client, direct_access_grants=direct_access_grants)
+    def run(self, name, clientId, redirectUris, protocol, publicClient, directAccessGrantsEnabled, skip_exists=True):
+        payload = {}
+        payload['name']=name
+        payload['clientId']=clientId
+        payload['redirectUris']=redirectUris
+        payload['protocol']=protocol
+        payload['publicClient']=publicClient
+        payload['directAccessGrantsEnabled']=directAccessGrantsEnabled
+
+        self.keycloak_admin.create_client(payload, skip_exists=skip_exists)