+++ CHANGELOG.md
@@ -9,3 +9,4 @@
  - Sync with official python-keycloak lib.
  - Rework get_group functions.
  - Add get_idps
+ - Add authentication flow actions